test_requirements.txt
-r requirements.txt
black==23.1.0
ipdb==0.13.11
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-mock==3.10.0